Update namespace of Liquid Custom Filters and Blocks
