--- lib/liquid/custom_filters/values.rb
@@ -1,7 +1,0 @@
-module Liquid
-  module CustomFilters
-    def values(list)
-      list.values
-    end
-  end
-end
