+++ lib/metanorma/plugin/datastruct/liquid/custom_filters/values.rb
@@ -1,0 +1,13 @@
+module Metanorma
+  module Plugin
+    module Datastruct
+      module Liquid
+        module CustomFilters
+          def values(list)
+            list.values
+          end
+        end
+      end
+    end
+  end
+end
